# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sales automation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List templates for a business
      parameters:
        - in: query
          name: businessId
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a template
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a template
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTemplateBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /templates/{id}/generate-ai:
    post:
      operationId: generateAiEmail
      tags: [templates]
      summary: Generate email content via AI for a specific lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAiEmailBody"
      responses:
        "200":
          description: Generated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedEmail"

  # ── Campaigns ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Template:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        name:
          type: string
        subjectTemplate:
          type: string
        htmlTemplate:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - businessId
        - name
        - subjectTemplate
        - htmlTemplate
        - createdAt
        - updatedAt
    CreateTemplateBody:
      type: object
      properties:
        businessId:
          type: integer
        name:
          type: string
        subjectTemplate:
          type: string
        htmlTemplate:
          type: string
      required:
        - businessId
        - name
        - subjectTemplate
        - htmlTemplate
    UpdateTemplateBody:
      type: object
      properties:
        name:
          type: string
        subjectTemplate:
          type: string
        htmlTemplate:
          type: string
    GenerateAiEmailBody:
      type: object
      properties:
        leadId:
          type: integer
      required:
        - leadId
    GeneratedEmail:
      type: object
      properties:
        subject:
          type: string
        html:
          type: string
      required:
        - subject
        - html

    # ── Campaign ────────────────────────────────────────────────────────────────
...